完善分词方法
