+++ hlp/mt/common/preprocess.py
@@ -9,14 +9,14 @@ sys.path.append('..')
 from sklearn.model_selection import train_test_split
 from model import trainer
 import config.get_config as _config
-import tensorflow_datasets as tfds
 from model import nmt_model
 from pathlib import Path
+from common import tokenize
 import tensorflow as tf
 import numpy
-import json
 import re
 import os
+import jieba
 
 
 def load_single_sentences(path, num_sentences, column):
@@ -85,6 +85,15 @@ def _preprocess_sentence_zh_char(sentence, start_word=_config.start_word, end_wo
     return s
 
 
+def _preprocess_sentence_zh_word(sentences, start_word=_config.start_word, end_word=_config.end_word):
+    """使用jieba进行分词前的预处理"""
+    sentences_word = []
+    for sentence in sentences:
+        sentence = start_word + ' ' + ' '.join(jieba.cut(sentence.strip())) + ' ' + end_word
+        sentences_word.append(sentence)
+    return sentences_word
+
+
 def _preprocess_sentences_zh(sentences, mode=_config.zh_tokenize_type, start_word=_config.start_word, end_word=_config.end_word):
     """
     对中文句子列表进行指定mode的预处理
@@ -93,8 +102,8 @@ def _preprocess_sentences_zh(sentences, mode=_config.zh_tokenize_type, start_wor
     if mode == 'CHAR':
         sentences = [_preprocess_sentence_zh_char(s, start_word, end_word) for s in sentences]
         return sentences
-    else:
-        return False
+    elif mode == 'WORD':
+        return _preprocess_sentence_zh_word(sentences)
 
 
 def preprocess_sentences(sentences, language):
@@ -105,267 +114,6 @@ def preprocess_sentences(sentences, language):
     elif language == "zh":
         mode = _config.zh_tokenize_type
         return _preprocess_sentences_zh(sentences, mode)
-
-
-def _create_tokenizer_bpe(sentences, save_path, start_word=_config.start_word
-                          , end_word=_config.end_word, target_vocab_size=_config.target_vocab_size):
-    """
-    根据指定语料生成字典
-    使用BPE分词
-    传入经预处理的句子
-    保存字典
-    Returns:字典，字典词汇量
-    """
-    tokenizer = tfds.features.text.SubwordTextEncoder.build_from_corpus(
-        sentences, target_vocab_size=target_vocab_size, reserved_tokens=[start_word, end_word])
-    tokenizer.save_to_file(save_path)
-    return tokenizer, tokenizer.vocab_size
-
-
-def _create_tokenizer_char(sentences, save_path):
-    """
-    根据指定语料生成字典
-    使用空格分词
-    传入经预处理的句子
-    保存字典
-    Returns:字典，字典词汇量
-    """
-    tokenizer = tf.keras.preprocessing.text.Tokenizer(filters='', oov_token='UNK')
-    tokenizer.fit_on_texts(sentences)
-    json_string = tokenizer.to_json()
-    with open(save_path, 'w') as f:
-        json.dump(json_string, f)
-    vocab_size = len(tokenizer.word_index)
-    return tokenizer, vocab_size
-
-
-# 根据指定模型生成及保存字典
-def create_tokenizer(sentences, language):
-    """
-    生成和保存指定语言的字典
-    所使用的模式为配置文件中设置的该语言的模式
-    支持的语言：en,zh
-    """
-    # 根据所选语言确定mode、save_path
-    if language == "en":
-        mode = _config.en_tokenize_type
-        save_path = _config.tokenizer_path_prefix+language+'_'+mode.lower()
-        if not os.path.exists(os.path.dirname(save_path)):
-            os.makedirs(os.path.dirname(save_path), exist_ok=True)
-    elif language == "zh":
-        mode = _config.zh_tokenize_type
-        save_path = _config.tokenizer_path_prefix+language+'_'+mode.lower()
-        if not os.path.exists(os.path.dirname(save_path)):
-            os.makedirs(os.path.dirname(save_path), exist_ok=True)
-
-    # 生成、保存目标语言字典
-    if mode == 'BPE':
-        return _create_tokenizer_bpe(sentences, save_path=save_path)
-    elif mode == 'CHAR':
-        return _create_tokenizer_char(sentences, save_path=save_path)
-
-
-def _get_tokenizer_bpe(path):
-    """从指定路径加载保存好的字典"""
-    tokenizer = tfds.features.text.SubwordTextEncoder.load_from_file(path)
-    return tokenizer, tokenizer.vocab_size
-
-
-def _get_tokenizer_char(path):
-    """从指定路径加载保存好的字典"""
-    with open(path) as f:
-        json_string = json.load(f)
-    tokenizer = tf.keras.preprocessing.text.tokenizer_from_json(json_string)
-    vocab_size = len(tokenizer.word_index)
-    return tokenizer, vocab_size
-
-
-# 取字典
-def get_tokenizer(language):
-    """
-    根据语言获取保存的字典
-    支持的语言：en、zh
-    """
-    # 根据所选语言确定mode、save_path
-    if language == "en":
-        mode = _config.en_tokenize_type
-        path = _config.tokenizer_path_prefix+language+'_'+mode.lower()
-    elif language == "zh":
-        mode = _config.zh_tokenize_type
-        path = _config.tokenizer_path_prefix+language+'_'+mode.lower()
-
-    if mode == 'BPE':
-        return _get_tokenizer_bpe(path)
-    elif mode == 'CHAR':
-        return _get_tokenizer_char(path)
-
-
-# 编码句子
-def _get_tokenized_tensor_bpe(sentences, tokenizer):
-    """
-    Args:
-        sentences: 需要编码的句子
-        tokenizer: 字典
-    Returns:已编码填充的句子及句子长度
-    """
-    sequences = [tokenizer.encode(s) for s in sentences]
-    sequences = tf.keras.preprocessing.sequence.pad_sequences(sequences, padding='post')
-    max_sequence_length = len(sequences[0])
-    return sequences, max_sequence_length
-
-
-def _encode_sentences_bpe(sentences, tokenizer):
-    """
-    Args:
-        sentences: 需要编码的句子列表
-        tokenizer: 字典
-
-    Returns:编码好的句子
-    """
-    sequences = [tokenizer.encode(s) for s in sentences]
-    sequences = tf.keras.preprocessing.sequence.pad_sequences(sequences, padding='post')
-    max_sequence_length = len(sequences[0])
-    return sequences, max_sequence_length
-
-
-def _encode_sentences_char(sentences, tokenizer):
-    """
-    Args:
-        sentences: 需要编码的句子列表
-        tokenizer: 字典
-
-    Returns:编码好的句子
-    """
-    sequences = tokenizer.texts_to_sequences(sentences)
-    sequences = tf.keras.preprocessing.sequence.pad_sequences(sequences, padding='post')
-    max_sequence_length = len(sequences[0])
-    return sequences, max_sequence_length
-
-
-def encode_sentences(sentences, tokenizer, language):
-    if language == "en":
-        mode = _config.en_tokenize_type
-    elif language == "zh":
-        mode = _config.zh_tokenize_type
-
-    if mode == 'BPE':
-        return _encode_sentences_bpe(sentences, tokenizer)
-    elif mode == 'CHAR':
-        return _encode_sentences_char(sentences, tokenizer)
-    else:
-        return False
-
-
-def get_start_token(start_word, tokenizer, language):
-    """
-    由于BPE分词的特殊性，在BPE获取start_token时需要在start_word后加一个空格
-    故在此使用方法对mode(编码模式)进行判断
-    返回 start_token  shape --> (1,) eg：[3]
-    """
-    if language == "en":
-        mode = _config.en_tokenize_type
-    elif language == "zh":
-        mode = _config.zh_tokenize_type
-
-    if mode == 'BPE':
-        start_word = start_word+' '
-        start_token, _ = _encode_sentences_bpe([start_word], tokenizer)
-    elif mode == 'CHAR':
-        start_token, _ = _encode_sentences_char([start_word], tokenizer)
-
-    start_token = [tf.squeeze(start_token)]
-    return start_token
-
-
-def _create_encoded_sentences_bpe(sentences, tokenizer, path):
-    """
-    将编码好的句子保存至文件，返回最大句子长度
-    Args:
-        sentences: 需要编码的句子
-        tokenizer: 字典
-        path:文件保存路径
-
-    Returns:最大句子长度
-    """
-    sequences = [tokenizer.encode(s) for s in sentences]
-    sequences = tf.keras.preprocessing.sequence.pad_sequences(sequences, padding='post')
-    max_sequence_length = len(sequences[0])
-    numpy.savetxt(path, sequences)
-    return max_sequence_length
-
-
-def _create_encoded_sentences_char(sentences, tokenizer, path):
-    """
-    将编码好的句子保存至文件，返回最大句子长度
-    Args:
-        sentences: 需要编码的句子
-        tokenizer: 字典
-
-    Returns:最大句子长度
-    """
-    sequences = tokenizer.texts_to_sequences(sentences)
-    sequences = tf.keras.preprocessing.sequence.pad_sequences(sequences, padding='post')
-    max_sequence_length = len(sequences[0])
-    numpy.savetxt(path, sequences)
-    return max_sequence_length
-
-
-def create_encoded_sentences(sentences, tokenizer, language):
-    """
-    根据所选语言将编码好的句子保存至文件，返回最大句子长度
-    Args:
-        sentences: 需要编码的句子
-        tokenizer: 字典
-        language: 语言类型 （en/zh）
-
-    Returns:最大句子长度
-    """
-    # 根据所选语言确定mode、save_path
-    if language == "en":
-        mode = _config.en_tokenize_type
-        save_path = _config.encoded_sequences_path_prefix+language
-        if not os.path.exists(os.path.dirname(save_path)):
-            os.makedirs(os.path.dirname(save_path), exist_ok=True)
-    elif language == "zh":
-        mode = _config.zh_tokenize_type
-        save_path = _config.encoded_sequences_path_prefix+language
-        if not os.path.exists(os.path.dirname(save_path)):
-            os.makedirs(os.path.dirname(save_path), exist_ok=True)
-
-    if mode == 'BPE':
-        return _create_encoded_sentences_bpe(sentences, tokenizer, save_path)
-    elif mode == 'CHAR':
-        return _create_encoded_sentences_char(sentences, tokenizer, save_path)
-    else:
-        return False
-
-
-def _decode_sentence_bpe(sequence, tokenizer):
-    return tokenizer.decode(sequence)
-
-
-def _decode_sentence_char(sequence, tokenizer):
-    sentence = [tokenizer.index_word[idx.numpy()] for idx in sequence
-                if idx != [tokenizer.word_index[_config.start_word]]]
-    sentence = ''.join(sentence)
-    return sentence
-
-
-def decode_sentence(sentences, tokenizer, language):
-    # 根据语言判断mode，支持语言：en、zh
-    if language == "en":
-        mode = _config.en_tokenize_type
-    elif language == "zh":
-        mode = _config.zh_tokenize_type
-    else:
-        print("暂不支持该语言："+language)
-
-    if mode == 'BPE':
-        return _decode_sentence_bpe(sentences, tokenizer)
-    elif mode == 'CHAR':
-        return _decode_sentence_char(sentences, tokenizer)
-    else:
-        return False
 
 
 def split_batch():
@@ -444,21 +192,23 @@ def train_preprocess():
 
     # 生成及保存字典
     print('正在生成、保存源语言(%s)字典(分词方式:%s)...' % (_config.source_lang, _config.en_tokenize_type))
-    tokenizer_source, vocab_size_source = create_tokenizer(sentences=source_sentences, language=_config.source_lang)
+    tokenizer_source, vocab_size_source = tokenize.create_tokenizer(sentences=source_sentences, language=_config.source_lang)
     print('生成英文字典大小:%d' % vocab_size_source)
     print('源语言字典生成、保存完毕！\n')
 
     print('正在生成、保存目标语言(%s)字典(分词方式:%s)...' % (_config.target_lang, _config.zh_tokenize_type))
-    tokenizer_target, vocab_size_target = create_tokenizer(sentences=target_sentences, language=_config.target_lang)
+    tokenizer_target, vocab_size_target = tokenize.create_tokenizer(sentences=target_sentences, language=_config.target_lang)
     print('生成目标语言字典大小:%d' % vocab_size_target)
     print('目标语言字典生成、保存完毕！\n')
 
     # 编码句子
     print("正在编码句子...")
-    max_sequence_length_source = create_encoded_sentences(sentences=source_sentences, tokenizer=tokenizer_source
-                                                          , language=_config.source_lang)
-    max_sequence_length_target = create_encoded_sentences(sentences=target_sentences, tokenizer=tokenizer_target
-                                                          , language=_config.target_lang)
+    max_sequence_length_source = tokenize.create_encoded_sentences(sentences=source_sentences
+                                                                   , tokenizer=tokenizer_source
+                                                                   , language=_config.source_lang)
+    max_sequence_length_target = tokenize.create_encoded_sentences(sentences=target_sentences
+                                                                   , tokenizer=tokenizer_target
+                                                                   , language=_config.target_lang)
     print('最大源语言(%s)句子长度:%d' % (_config.source_lang, max_sequence_length_source))
     print('最大目标语言(%s)句子长度:%d' % (_config.target_lang, max_sequence_length_target))
     print("句子编码完毕！\n")
@@ -472,13 +222,13 @@ def load_model():
     """
     # 加载源语言字典
     print("正在加载源语言(%s)字典..." % _config.source_lang)
-    tokenizer_source, vocab_size_source = get_tokenizer(language=_config.source_lang)
+    tokenizer_source, vocab_size_source = tokenize.get_tokenizer(language=_config.source_lang)
     print('源语言字典大小:%d' % vocab_size_source)
     print('源语言字典加载完毕！\n')
 
     # 加载目标语言字典
     print("正在加载目标语言(%s)字典..." % _config.target_lang)
-    tokenizer_target, vocab_size_target = get_tokenizer(language=_config.target_lang)
+    tokenizer_target, vocab_size_target = tokenize.get_tokenizer(language=_config.target_lang)
     print('目标语言字典大小:%d' % vocab_size_target)
     print('目标语言字典加载完毕！\n')
 
@@ -511,10 +261,10 @@ def main():
     模块方法测试
     """
     # 加载中英文字典
-    tokenizer_en, vocab_size_en = get_tokenizer(path="../data/en_tokenizer"
-                                                , mode=_config.en_tokenize_type)
-    tokenizer_ch, vocab_size_ch = get_tokenizer(path='../data/tokenizer/ch_tokenizer.json'
-                                                , mode=_config.ch_tokenize_type)
+    tokenizer_en, vocab_size_en = tokenize.get_tokenizer(path="../data/en_tokenizer"
+                                                         , mode=_config.en_tokenize_type)
+    tokenizer_ch, vocab_size_ch = tokenize.get_tokenizer(path='../data/tokenizer/ch_tokenizer.json'
+                                                         , mode=_config.ch_tokenize_type)
     print(vocab_size_en)
     print(vocab_size_ch)
     en = 'Transformer is good.'
@@ -529,9 +279,9 @@ def main():
 
     # 编码句子
     print("编码后的句子")
-    en, _ = encode_sentences(en, tokenizer_en, mode='BPE')
+    en, _ = tokenize.encode_sentences(en, tokenizer_en, mode='BPE')
 
-    ch, _ = encode_sentences(ch, tokenizer_ch, mode='WORD')
+    ch, _ = tokenize.encode_sentences(ch, tokenizer_ch, mode='WORD')
     print(en)
     for ts in en[0]:
         print('{} ----> {}'.format(ts, tokenizer_en.decode([ts])))
